nop bai tap

+++ thymeleaf_product/src/main/java/com/example/thymeleaf_product/entity/Product.java
@@ -3,14 +3,14 @@ package com.example.thymeleaf_product.entity;
 public class Product {
     private int id;
     private String name;
-    private double price;
+    private int price;
     private String description;
     private String producer;
 
     public Product() {
     }
 
-    public Product(int id, String name, double price, String description, String producer) {
+    public Product(int id, String name, int price, String description, String producer) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -34,11 +34,11 @@ public class Product {
         this.name = name;
     }
 
-    public double getPrice() {
+    public int getPrice() {
         return price;
     }
 
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
